split test_io.py out from test_utils.py
