+++ tests/test_utils.py
@@ -1,15 +1,12 @@
 from __future__ import annotations
 
-import os
 from copy import deepcopy
 from datetime import datetime
-from typing import TYPE_CHECKING, Any, Sequence
-from unittest.mock import patch
+from typing import Any, Sequence
 
 import pandas as pd
 import plotly.graph_objects as go
 import pytest
-from matplotlib import pyplot as plt
 from matplotlib.offsetbox import AnchoredText
 
 from pymatviz.utils import (
@@ -18,16 +15,10 @@ from pymatviz.utils import (
     annotate_metrics,
     bin_df_cols,
     df_to_arrays,
-    df_to_pdf,
     get_crystal_sys,
     patch_dict,
-    save_fig,
 )
 from tests.conftest import y_pred, y_true
-
-
-if TYPE_CHECKING:
-    from pathlib import Path
 
 
 @pytest.mark.parametrize(
@@ -141,51 +132,6 @@ def test_df_to_arrays_strict(strict: bool) -> None:
             assert args == ("foo", "bar")
 
 
-@pytest.mark.parametrize("fig", [go.Figure(), plt.figure()])
-@pytest.mark.parametrize("ext", ["html", "svelte", "png", "svg", "pdf"])
-@pytest.mark.parametrize(
-    "plotly_config", [None, {"showTips": True}, {"scrollZoom": True}]
-)
-@pytest.mark.parametrize("env_disable", [[], ["CI"]])
-@patch.dict(os.environ, {"CI": "1"})
-def test_save_fig(
-    fig: go.Figure | plt.Figure | plt.Axes,
-    ext: str,
-    tmp_path: Path,
-    plotly_config: dict[str, Any] | None,
-    env_disable: list[str],
-) -> None:
-    if isinstance(fig, plt.Figure) and ext in ("svelte", "html"):
-        pytest.skip("saving to Svelte file not supported for matplotlib figures")
-
-    path = f"{tmp_path}/fig.{ext}"
-    save_fig(fig, path, plotly_config=plotly_config, env_disable=env_disable)
-
-    if any(var in os.environ for var in env_disable):
-        # if CI env var is set, we should not save the figure
-        assert not os.path.exists(path)
-        return
-
-    assert os.path.isfile(path)
-
-    if ext in ("svelte", "html"):
-        with open(path) as file:
-            html = file.read()
-        if plotly_config and plotly_config.get("showTips"):
-            assert '"showTips": true' in html
-        else:
-            assert '"showTips": false' in html
-        assert '"modeBarButtonsToRemove": ' in html
-        assert '"displaylogo": false' in html
-        if plotly_config and plotly_config.get("scrollZoom"):
-            assert '"scrollZoom": true' in html
-
-        if ext == "svelte":
-            assert html.startswith("<div {...$$props}>")
-        else:
-            assert html.startswith("<div>")
-
-
 @pytest.mark.parametrize(
     "bin_by_cols, group_by_cols, n_bins, expected_n_bins, "
     "verbose, kde_col, expected_n_rows",
@@ -247,23 +193,6 @@ def test_bin_df_cols_raises_value_error() -> None:
         bin_df_cols(df, bin_by_cols, n_bins=[2])
 
     assert "len(bin_by_cols)=2 != len(n_bins)=1" in str(exc.value)
-
-
-def test_plotly_pdf_no_mathjax_loading(tmp_path: Path) -> None:
-    # https://github.com/plotly/plotly.py/issues/3469
-    PyPDF2 = pytest.importorskip("PyPDF2")
-
-    fig = go.Figure()
-    fig.add_trace(go.Scatter(x=[1, 2], y=[3, 4]))
-    path = f"{tmp_path}/test.pdf"
-    save_fig(fig, path)
-
-    # check PDF doesn't contain "Loading [MathJax]/extensions/MathMenu.js"
-    with open(path, "rb") as f:
-        pdf = PyPDF2.PdfReader(f)
-        assert len(pdf.pages) == 1
-        text = pdf.pages[0].extract_text()
-        assert "Loading [MathJax]/extensions/MathMenu.js" not in text
 
 
 sample_dict = {"a": 1, "b": None, "c": [3, 4]}
@@ -334,38 +263,3 @@ def test_patch_dict_remove_key_inside_context() -> None:
 
 
 assert ref_sample_dict == sample_dict
-
-
-@pytest.mark.parametrize(
-    "crop, size, style",
-    [
-        # test with cropping, default size, and no extra style
-        (True, "landscape", ""),
-        # test without cropping, portrait size, and additional styles
-        (False, "portrait", "body { margin: 0; padding: 1em; }"),
-    ],
-)
-def test_df_to_pdf(crop: bool, size: str, style: str, tmp_path: Path) -> None:
-    # Create a test DataFrame and Styler object
-    df: pd.DataFrame = pd._testing.makeDataFrame()  # random data
-    file_path = tmp_path / "test.pdf"
-
-    # Execute the function
-    df_to_pdf(df.style, file_path, crop=crop, size=size, style=style)
-
-    # Check if the file is created
-    assert file_path.is_file()
-
-    with open(file_path, "rb") as pdf_file:
-        contents = pdf_file.read()
-
-    # TODO: Add more specific checks here, like file content validation
-    assert contents[:4] == b"%PDF"
-
-    # Test file overwrite behavior
-    file_size_before = file_path.stat().st_size
-    df_to_pdf(df.style, file_path, crop=crop, size=size, style=style)
-    file_size_after = file_path.stat().st_size
-
-    # file size should be the same since content is unchanged
-    assert file_size_before - 10 <= file_size_after <= file_size_before + 10
